pyproject.toml
[project]
name = "personal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.15",
    "llama-index>=0.12.52",
    "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-vector-stores-chroma>=0.4.2",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "plotly>=6.2.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-magic>=0.4.27",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.47.1",
    "torch>=2.7.1",
    "transformers>=4.54.0",
]
